Refactor: 하나의 파일에 너무 많은게 들어있어서 분리해줬습니다 . .
